Preventing deletion of Users and Posts
- Admins can delete all posts
- Users can only delete their own posts

+++ controller/post.js
@@ -54,10 +54,9 @@ exports.create_post = [
 
     await post.save();
 
-    return res.json(post);
+    return res.json({message: "Post Created!", post: post});
   }),
 ];
-
 exports.delete_post = asyncHandler(async (req, res, next) => {
   if (req.body.postId == null || req.body.postId == "") {
     const err = new Error();
@@ -73,10 +72,15 @@ exports.delete_post = asyncHandler(async (req, res, next) => {
     err.status = 404;
     err.message = "Post not found!";
     return res.json(err);
-  } else {
-    await Post.findByIdAndDelete(req.body.postId);
-    return res.json({ message: "Post was deleted successfully!" });
+  } 
+
+  if (post.author.toString() !== req.user._id.toString() && !req.user.admin) {
+    return res.status(403).json({ message: "You are not authorized to delete this post." });
   }
+   
+  await Post.findByIdAndDelete(req.body.postId);
+  return res.json({ message: "Post was deleted successfully!" });
+  
 });
 exports.get_post_details = asyncHandler(async (req, res, next) => {
   const post = await Post.findById(req.params.id).populate("author").exec();
@@ -125,6 +129,10 @@ exports.update_post = [
       return res.status(404).json("Post not found!");
     }
 
+    if (post.author.toString() !== req.user._id.toString()) {
+      return res.status(403).json({ message: "You are not authorized to update this post." });
+    }
+
     const updatedPost = new Post({
       _id: req.params.id,
       author: post.author._id,
@@ -136,6 +144,6 @@ exports.update_post = [
 
     await Post.findByIdAndUpdate(req.params.id, updatedPost, {});
 
-    return res.json(updatedPost);
+    return res.json({ message: "Updated Post!", post: updatedPost });
   }),
 ];
